refactor(abstractions): migrate pagination types to int
Transition pagination index and size from uint to int for CLS
compliance and better alignment with EF Core and database providers.
Add guard clauses for index and size validation, and simplify
arithmetic in skip calculation. Update existing test suites to
ensure full compatibility with the new signed integer types.

+++ src/NFramework.Persistence.EFCore/Extensions/PaginationExtensions.cs
@@ -42,7 +42,7 @@ public static class PaginationExtensions
 
             List<T> items = await source
                 .Skip((int)skipItems)
-                .Take((int)paging.Size)
+                .Take(paging.Size)
                 .ToListAsync(cancellationToken)
                 .ConfigureAwait(false);
